separate FeedMapVC extensions
