+++ wmsy/Pages/FeedMapPage/FeedMapVC+Filter.swift
@@ -1,0 +1,37 @@
+//
+//  FeedMapVC+Filter.swift
+//  wmsy
+//
+//  Created by C4Q on 3/27/18.
+//  Copyright © 2018 C4Q. All rights reserved.
+//
+
+import UIKit
+import SnapKit
+import GoogleMaps
+import SVProgressHUD
+
+extension FeedMapVC: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        let cell = collectionView.cellForItem(at: indexPath) as! WhimCategoryCollectionViewCell
+        let tuple = categoryList[indexPath.row]
+        var selectedCategory = tuple
+        self.filtersView.categoryLabel.text = "Filter Whims by: \(selectedCategory.0)"
+        DBService.manager.getCategoryWhims(fromCategory: selectedCategory.0) { (whims) in
+            self.feedWhims = whims
+        }
+    }
+}
+
+extension FeedMapVC: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return categoryList.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "FilterCategoryCell", for: indexPath) as! WhimCategoryCollectionViewCell
+        let categoryImage = categoryList[indexPath.row].1
+        cell.categoryImage.image = categoryImage
+        return cell
+    }
+}
