+++ wmsy/Pages/FeedMapPage/FeedMapVC.swift
@@ -11,9 +11,6 @@ import SnapKit
 import GoogleMaps
 import SVProgressHUD
 
-//protocol ParentDelegate: class {
-//    func updateChildren(whims: [Whim]) -> Void
-//}
 
 class FeedMapVC: MenuedViewController {
     
@@ -32,7 +29,6 @@ class FeedMapVC: MenuedViewController {
             print("userLocation set")
             DBService.manager.getClosestWhims(location: userLocation) { (whims) in
                 self.feedWhims = whims
-                //                self.delegate?.updateChildren(whims: whims)
             }
         }
     }
@@ -58,20 +54,16 @@ class FeedMapVC: MenuedViewController {
             }
         }
     }
-
     
     
     override func viewDidLoad() {
         super.viewDidLoad()
         SVProgressHUD.dismiss()
         
-        filtersView.isHidden = true
-        mapView.isHidden = true
-        
-        filtersView.categoriesCV.delegate = self
-        filtersView.categoriesCV.dataSource = self
-        filtersView.categoriesCV.reloadData()
-        filtersView.clearSearchButton.addTarget(self, action: #selector(clearSearch), for: .touchUpInside)
+        filtersView.isHidden = false
+        mapView.isHidden = false
+//        filtersView.isHidden = true
+//        mapView.isHidden = true
         
         let mylocation = mapView.mapView.myLocation
         mapView.mapView.camera = GMSCameraPosition.camera(withLatitude: (mylocation?.coordinate.latitude)!,
@@ -80,18 +72,20 @@ class FeedMapVC: MenuedViewController {
         mapView.mapView.settings.myLocationButton = true
         mapView.mapView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
         
-        
         locationManager = CLLocationManager()
         locationManager.desiredAccuracy = kCLLocationAccuracyBest
         locationManager.requestAlwaysAuthorization()
         locationManager.distanceFilter = 50
         locationManager.startUpdatingLocation()
         self.locationManager.delegate = self
-
         
         view.addSubview(feedView)
         feedView.snp.makeConstraints { (make) in
-            make.edges.equalTo(view.safeAreaLayoutGuide)
+//            make.edges.equalTo(view.safeAreaLayoutGuide)
+            make.top.equalTo(view.safeAreaLayoutGuide.snp.top)
+            make.leading.equalTo(view.safeAreaLayoutGuide.snp.leading)
+            make.trailing.equalTo(view.safeAreaLayoutGuide.snp.trailing)
+            make.height.equalTo(view.safeAreaLayoutGuide.snp.height).multipliedBy(0.84)
         }
         
         feedView.tableView.dataSource = self
@@ -104,13 +98,18 @@ class FeedMapVC: MenuedViewController {
         filtersView.snp.makeConstraints { (make) in
             make.leading.equalTo(view.safeAreaLayoutGuide.snp.leading)
             make.trailing.equalTo(view.safeAreaLayoutGuide.snp.trailing)
-            //            make.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom).offset(-50)
+            make.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom)
             //            make.top.equalTo(view.safeAreaLayoutGuide.snp.bottom).offset(-200)
             //                        make.top.equalTo(view.safeAreaLayoutGuide.snp.top).offset(50)
-            make.top.equalTo(view.safeAreaLayoutGuide.snp.top)
+//            make.top.equalTo(view.safeAreaLayoutGuide.snp.top)
             //            make.height.equalTo(view.safeAreaLayoutGuide.snp.height).multipliedBy(0.25)
             make.height.equalTo(view.safeAreaLayoutGuide.snp.height).multipliedBy(0.16)
         }
+        
+        filtersView.categoriesCV.delegate = self
+        filtersView.categoriesCV.dataSource = self
+        filtersView.categoriesCV.reloadData()
+        filtersView.clearSearchButton.addTarget(self, action: #selector(clearSearch), for: .touchUpInside)
         
         view.addSubview(mapView)
         mapView.snp.makeConstraints { (make) in
@@ -119,11 +118,7 @@ class FeedMapVC: MenuedViewController {
             make.top.equalTo(filtersView.snp.bottom)
             make.height.equalTo(view.safeAreaLayoutGuide.snp.height).multipliedBy(0.84)
         }
-        
-        
-
         configureNavBar()
-
     }
     
     
@@ -141,9 +136,6 @@ class FeedMapVC: MenuedViewController {
         
     }
     
-    @objc func clearSearch() {
-        print("need to add functionality to clear the search category")
-    }
     
     @objc func hostAWhim() {
         navigationController?.pushViewController(CreateWhimTVC(), animated: true)
@@ -162,67 +154,70 @@ class FeedMapVC: MenuedViewController {
         let topRightBarItem = UIBarButtonItem(image: #imageLiteral(resourceName: "mapIcon"), style: .plain, target: self, action: #selector(showMap))
         navigationItem.rightBarButtonItem = topRightBarItem
     }
-}
 
-
-
-extension FeedMapVC: GMSMapViewDelegate{
-    func mapView(_ mapView: GMSMapView, didTap marker: GMSMarker) -> Bool {
-        let camera = GMSCameraPosition.camera(withLatitude: marker.position.latitude,
-                                             longitude: marker.position.longitude,
-                                             zoom: 15.0)
-        self.mapView.mapView.animate(to: camera)
-        let dict = marker.userData as? [String: String]
-        self.mapView.detailView.whimTitle.text = dict!["title"]
-        self.mapView.detailView.whimDescription.text = dict!["description"]
-        let hostURL = URL(string: dict!["hostImageURL"]!)
-        let hostID = dict!["hostID"]
-        DBService.manager.getAppUser(with: hostID!) { (appUser) in
-           self.currentUser = appUser
-        }
-        self.mapView.detailView.userPicture.kf.setImage(with: hostURL, for: .normal)
-        self.mapView.detailView.isHidden = false
-
-        return true
-    }
-
-    func mapView(_ mapView: GMSMapView, didTapAt coordinate: CLLocationCoordinate2D) {
-        self.mapView.detailView.isHidden = true
-    }
-
-}
-
-extension FeedMapVC: CLLocationManagerDelegate{
-    
-    // Handle incoming location events.
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        let location: CLLocation = locations.last!
-        print("Location: \(location)")
-        self.userLocation = location
-        
-    }
-    
-    // Handle authorization for the location manager.
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        switch status {
-        case .restricted:
-            print("Location access was restricted.")
-        case .denied:
-            print("User denied access to location.")
-        case .notDetermined:
-            print("Location status not determined.")
-        case .authorizedAlways: fallthrough
-        case .authorizedWhenInUse:
-            print("Location status is OK.")
-        }
-    }
-    
-    // Handle location manager errors.
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        self.locationManager.stopUpdatingLocation()
-        print("Error: \(error)")
+    @objc func clearSearch() {
+        print("need to add functionality to clear the search category")
     }
 }
+
+
+
+//extension FeedMapVC: GMSMapViewDelegate{
+//    func mapView(_ mapView: GMSMapView, didTap marker: GMSMarker) -> Bool {
+//        let camera = GMSCameraPosition.camera(withLatitude: marker.position.latitude,
+//                                             longitude: marker.position.longitude,
+//                                             zoom: 15.0)
+//        self.mapView.mapView.animate(to: camera)
+//        let dict = marker.userData as? [String: String]
+//        self.mapView.detailView.whimTitle.text = dict!["title"]
+//        self.mapView.detailView.whimDescription.text = dict!["description"]
+//        let hostURL = URL(string: dict!["hostImageURL"]!)
+//        let hostID = dict!["hostID"]
+//        DBService.manager.getAppUser(with: hostID!) { (appUser) in
+//           self.currentUser = appUser
+//        }
+//        self.mapView.detailView.userPicture.kf.setImage(with: hostURL, for: .normal)
+//        self.mapView.detailView.isHidden = false
+//
+//        return true
+//    }
+//
+//    func mapView(_ mapView: GMSMapView, didTapAt coordinate: CLLocationCoordinate2D) {
+//        self.mapView.detailView.isHidden = true
+//    }
+//}
+
+//extension FeedMapVC: CLLocationManagerDelegate{
+//
+//    // Handle incoming location events.
+//    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+//        let location: CLLocation = locations.last!
+//        print("Location: \(location)")
+//        self.userLocation = location
+//
+//    }
+//
+//    // Handle authorization for the location manager.
+//    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
+//        switch status {
+//        case .restricted:
+//            print("Location access was restricted.")
+//        case .denied:
+//            print("User denied access to location.")
+//        case .notDetermined:
+//            print("Location status not determined.")
+//        case .authorizedAlways: fallthrough
+//        case .authorizedWhenInUse:
+//            print("Location status is OK.")
+//        }
+//    }
+//
+//    // Handle location manager errors.
+//    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
+//        self.locationManager.stopUpdatingLocation()
+//        print("Error: \(error)")
+//    }
+//}
 
 
 
@@ -269,43 +264,41 @@ extension FeedMapVC: CLLocationManagerDelegate{
 //    }
 //}
 
-extension FeedMapVC: mapDetailViewDelegate {
-    func interestPressed() {
-        print("interest is being pressed")
-    }
-    
-    func userPicturePressed() {
-        present(GuestProfileVC(), animated: true, completion: nil)
-    }
-    
-    
-}
+//extension FeedMapVC: mapDetailViewDelegate {
+//    func interestPressed() {
+//        print("interest is being pressed")
+//    }
+//
+//    func userPicturePressed() {
+//        present(GuestProfileVC(), animated: true, completion: nil)
+//    }
+//}
 
 
-extension FeedMapVC: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let cell = collectionView.cellForItem(at: indexPath) as! WhimCategoryCollectionViewCell
-        let tuple = categoryList[indexPath.row]
-        var selectedCategory = tuple
-        self.filtersView.categoryLabel.text = "Filter Whims by: \(selectedCategory.0)"
-        DBService.manager.getCategoryWhims(fromCategory: selectedCategory.0) { (whims) in
-            self.feedWhims = whims
-        }
-    }
-}
-
-extension FeedMapVC: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return categoryList.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "FilterCategoryCell", for: indexPath) as! WhimCategoryCollectionViewCell
-        let categoryImage = categoryList[indexPath.row].1
-        cell.categoryImage.image = categoryImage
-        return cell
-    }
-}
+//extension FeedMapVC: UICollectionViewDelegateFlowLayout {
+//    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+//        let cell = collectionView.cellForItem(at: indexPath) as! WhimCategoryCollectionViewCell
+//        let tuple = categoryList[indexPath.row]
+//        var selectedCategory = tuple
+//        self.filtersView.categoryLabel.text = "Filter Whims by: \(selectedCategory.0)"
+//        DBService.manager.getCategoryWhims(fromCategory: selectedCategory.0) { (whims) in
+//            self.feedWhims = whims
+//        }
+//    }
+//}
+//
+//extension FeedMapVC: UICollectionViewDataSource {
+//    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+//        return categoryList.count
+//    }
+//
+//    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+//        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "FilterCategoryCell", for: indexPath) as! WhimCategoryCollectionViewCell
+//        let categoryImage = categoryList[indexPath.row].1
+//        cell.categoryImage.image = categoryImage
+//        return cell
+//    }
+//}
